works for the popup, doesnt for the content script
